pyproject.toml
[project]
name = "project-dashboard"
version = "0.1.0"
description = "\"Project management Api\""
authors = [
    {name = "Kovacs_Roland_Gergo",email = "kovaxrawland@mailbox.unideb.hu"}
]
packages=[{include="app", from="src" }]
license = {text = "MIT Licence"}
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "boto3 (>=1.40.13,<2.0.0)",
    "moto (>=5.1.10,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
